// src/types/formularioDeAvaliacaoCursoIntrodutorioMindfulness.d.ts
export interface FormularioDeAvaliacaoCursoIntrodutorioMindfulnessI {
  /** Auto-incrementing primary key */
  id: number;

  createdAt?: string;

  fullName?: string;
  birthDate?: string;
  satisfactionLevel?: string;

  likedMostAndHighlights?: string;
  likedLeastSuggestions?: string;
  personalChange?: string;
  selfDifference?: string;
  knowledgeMomentsImportance?: string;
  tellAFriend?: string;
  benefits?: string;
  authorizeUse?: string;
}